fix: sparkle skill default off on self, fix hot reload crash
