+++ src/components/optimizerTab/optimizerForm/StatSimulationDisplay.tsx
@@ -12,15 +12,9 @@ import {
   saveStatSimulationBuildFromForm,
   startOptimizerStatSimulation
 } from "lib/statSimulationController.tsx";
-import { Parts } from "lib/constants";
+import { BodyStatOptions, FeetStatOptions, LinkRopeStatOptions, Parts, PlanarSphereStatOptions } from "lib/constants";
 import { Assets } from "lib/assets";
-import {
-  BodyStatOptions,
-  FeetStatOptions,
-  LinkRopeStatOptions,
-  OrnamentSetTagRenderer,
-  PlanarSphereStatOptions
-} from "components/optimizerTab/optimizerForm/RelicMainSetFilters";
+import { OrnamentSetTagRenderer, } from "components/optimizerTab/optimizerForm/RelicMainSetFilters";
 import GenerateOrnamentsOptions from "components/optimizerTab/optimizerForm/OrnamentsOptions";
 import { GenerateBasicSetsOptions } from "components/optimizerTab/optimizerForm/SetsOptions";
 import { Utils } from "lib/utils";
